Slightly refactors the cli
