+++ packages/yarnpkg-builder/types/yarnVersion.d.ts
@@ -1,0 +1,1 @@
+declare const YARN_VERSION: string;
