+++ packages/yarnpkg-cli/sources/index.ts
@@ -1,3 +1,4 @@
 export {BaseCommand}            from './BaseCommand';
 export {WorkspaceRequiredError} from './WorkspaceRequiredError';
+export {main}                   from './main';
 export {openWorkspace}          from './openWorkspace';
