+++ packages/yarnpkg-cli/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "outDir": "lib",
+    "declaration": true
+  },
+  "include": [
+    "sources/**/*",
+    "../*/types/**/*"
+  ]
+}
